Update .NET Core version in GitHub Actions workflow
Updated the .NET Core version used in the `main_mycontosocrafts.yml` GitHub Actions workflow file from '6.0.x' to '8.0.x'. This change is reflected in the `dotnet-version` parameter of the `actions/setup-dotnet@v1` action. The `include-prerelease: true` line has been added to consider prerelease versions of .NET Core. This update will affect the .NET Core setup step and all subsequent steps that rely on it, including the 'Restore dependencies' and 'Build with dotnet' steps.